LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/clear-prefetch-cache-after-clear-site-data-cache.https-expected.txt


FAIL clear-site-data prefetchCache headers should prevent it from being fetched assert_equals: expected (undefined) undefined but got (string) "prefetch"
PASS clear-site-data cache headers should prevent it from being fetched
PASS clear-site-data prefetchCache in same window should clear its prefetch cache
